Rejiggered capabilities setup and lazy-load events
The plugin was initially configured to lazy-load only during the 'InsertEnter' event to speed up startup. This approach led to issues with completion and language server functionalities when immediately entering command mode. To resolve this, the plugin is now set to load on the 'CmdlineEnter' event as well. I still don't know how to properly deal with capabilities.